[Add] 자신의 게시글만 수정, 삭제할 수 있도록 권한 검증 추가

+++ src/main/java/com/ormi5/movieblog/postcontroller/PostController.java
@@ -3,24 +3,17 @@ package com.ormi5.movieblog.postcontroller;
 import com.ormi5.movieblog.post.PostDto;
 import com.ormi5.movieblog.post.PostResponseDto;
 
-import com.ormi5.movieblog.comment.CommentDto;
-import com.ormi5.movieblog.commentcontroller.CommentService;
-import com.ormi5.movieblog.post.Post;
-import com.ormi5.movieblog.post.PostDto;
-
-import com.ormi5.movieblog.post.PostResponseDto;
 import com.ormi5.movieblog.user.User;
+import com.ormi5.movieblog.usercontroller.UserService;
+
 import lombok.RequiredArgsConstructor;
 
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
 
-import java.util.Map;
-import java.util.Optional;
+import java.security.Principal;
 import java.util.List;
 
 @Controller
@@ -28,6 +21,7 @@ import java.util.List;
 @RequiredArgsConstructor
 public class PostController {
 	private final PostService postService;
+	private final UserService userService;
 
 	/**
 	 * 게시글 생성 메서드
@@ -35,12 +29,11 @@ public class PostController {
 	 * @author yuseok
 	 * @param postDto 생성할 게시글 정보가 담긴 DTO
 	 * @return 생성된 게시글 정보가 담긴 DTO
-	@PostMapping
-	public ResponseEntity<PostDto> createPost(@RequestBody PostDto postDto) {
-		PostDto createdPost = postService.createPost(postDto);
+	 @PostMapping public ResponseEntity<PostDto> createPost(@RequestBody PostDto postDto) {
+	 PostDto createdPost = postService.createPost(postDto);
 
-		return new ResponseEntity<>(createdPost, HttpStatus.CREATED);
-	}*/
+	 return new ResponseEntity<>(createdPost, HttpStatus.CREATED);
+	 }*/
 
 	// 모든 게시글을 조회하는 메서드
 	@GetMapping("/list")
@@ -146,8 +139,16 @@ public class PostController {
 	 * @author nayoung
 	 */
 	@GetMapping("/{postId}/edit")
-	public String editForm(@PathVariable("postId") Integer postId, Model model) {
+	public String editForm(@PathVariable("postId") Integer postId, Model model, Principal principal) {
 		PostDto post = postService.getPostById(postId);
+
+		User user = userService.findByUsername(principal.getName());
+
+		if (!post.getUser().getId().equals(user.getId())) {
+			System.out.println("수정 불가!");
+			return "redirect:/posts/{postId}";
+		}
+
 		model.addAttribute("post", post);
 		return "post/edit";
 	}
@@ -188,7 +189,16 @@ public class PostController {
 	 * @return result ResponseEntity
 	 */
 	@PostMapping("/{id}/delete")
-	public String deletePostById(@PathVariable("id") Integer id) {
+	public String deletePostById(@PathVariable("id") Integer id, Principal principal) {
+		PostDto postDto = postService.getPostById(id);
+
+		User user = userService.findByUsername(principal.getName());
+
+		if (!postDto.getUser().getId().equals(user.getId())) {
+			System.out.println("삭제 불가!");
+			return "redirect:/posts/" + postDto.getPostId();
+		}
+
 		boolean deleted = postService.deletePost(id);
 		return "redirect:/board";
 	}
